Keep integration generator helpers easier to extend
Extract the reusable parsing and tool-template helpers from the
integration generator so the command file stays focused on prompting and
file generation orchestration.

Constraint: Preserve integration generator behavior and existing scaffold content
Constraint: Repo pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Refactoring OAuth route template generation in the same pass | broader than one queue item needs
Confidence: high
Scope-risk: narrow
Directive: Keep integration-generator-helpers.ts limited to pure parsing/template helpers; leave file-writing orchestration in integration-generator.ts
Tested: deno test --no-check --allow-all cli/commands/generate/integration-generator.test.ts; deno fmt --check cli/commands/generate/integration-generator.ts cli/commands/generate/integration-generator-helpers.ts cli/commands/generate/integration-generator.test.ts; deno lint cli/commands/generate/integration-generator.ts cli/commands/generate/integration-generator-helpers.ts cli/commands/generate/integration-generator.test.ts; git diff --check
Not-tested: Pre-commit verify:quick (fails on existing CLI boundary violations in cli/commands/extension/*.ts)
